Updating serialization-related test classes
